workspace: fix clippy 1.98 fallout outside core
Same story as the core commit, for the rest of the workspace: clippy
1.98 collapses nested ifs more aggressively (87 of the findings), and
the remainder is the familiar set of map_or defaults with function
calls, unwraps right after variant checks, explicit loop counters,
sort_by with a key comparator, slice-from-ref clones, as_chunks for
constant-size chunking, and format widths that never did anything.

rustc 1.98 also stopped counting derive expansions as uses, which
exposed two genuinely unconstructed types (the MCP InitializeRequest
payload shape and the simulator's QueryTypes); both keep their
definitions under an explicit allow(dead_code) with a note instead of
being deleted in a toolchain change. The parser's large_stack_frames
expectation no longer fires on 1.98 and is removed.

Tests: cargo clippy --workspace --all-features --all-targets is clean
under 1.98; cargo fmt --check clean.

+++ bindings/javascript/src/browser.rs
@@ -188,7 +188,7 @@ impl File for OpfsFile {
         let buffer = buffer.as_mut_slice();
         if web_worker {
             let result = unsafe { read(handle, buffer.as_mut_ptr(), buffer.len(), pos as i32) };
-            c.complete(result as i32);
+            c.complete(result);
         } else {
             let completion_no = self.opfs.register_completion(c.clone());
             unsafe {
@@ -224,7 +224,7 @@ impl File for OpfsFile {
         let buffer = buffer.as_slice();
         if web_worker {
             let result = unsafe { write(handle, buffer.as_ptr(), buffer.len(), pos as i32) };
-            c.complete(result as i32);
+            c.complete(result);
         } else {
             let completion_no = self.opfs.register_completion(c.clone());
             unsafe {
@@ -250,7 +250,7 @@ impl File for OpfsFile {
         let handle = self.handle;
         if web_worker {
             let result = unsafe { sync(handle) };
-            c.complete(result as i32);
+            c.complete(result);
         } else {
             let completion_no = self.opfs.register_completion(c.clone());
             unsafe { sync_async(handle, completion_no) };
@@ -273,7 +273,7 @@ impl File for OpfsFile {
         let handle = self.handle;
         if web_worker {
             let result = unsafe { truncate(handle, len as usize) };
-            c.complete(result as i32);
+            c.complete(result);
         } else {
             let completion_no = self.opfs.register_completion(c.clone());
             unsafe { truncate_async(handle, len as usize, completion_no) };
